bug fixes from system testing.

+++ dnx_secmods/cfirewall/fw_main.pxd
@@ -155,14 +155,14 @@ cdef struct FWrule:
     # source
     bint      enabled # enable or disable rule from being matched against
     u_int8_t  s_zone
-    u_int32_t s_net_id
+    long      s_net_id
     u_int32_t s_net_mask
     u_int32_t s_port_start # extra 16 bits to include protocol. bitwise when eval tcp/80(webui) > 6/80
     u_int16_t s_port_end
 
     # destination
     u_int8_t  d_zone
-    long d_net_id
+    long      d_net_id
     u_int32_t d_net_mask
     u_int32_t d_port_start # extra 16 bits to include protocol. bitwise when eval tcp/80(webui) > 6/80
     u_int16_t d_port_end
